+++ pol-core/pol/module/polsystemmod.cpp
@@ -16,34 +16,38 @@
 #include <fstream>
 #include <string>
 
-#include "../../bscript/berror.h"
-#include "../../bscript/bobject.h"
-#include "../../bscript/dict.h"
-#include "../../bscript/impstr.h"
-#include "../../clib/clib.h"
-#include "../../clib/clib_MD5.h"
-#include "../../clib/fileutil.h"
-#include "../../clib/rawtypes.h"
-#include "../../clib/strutil.h"
-#include "../../clib/threadhelp.h"
-#include "../../plib/pkg.h"
-#include "../../plib/systemstate.h"
-#include "../cmdlevel.h"
-#include "../core.h"
-#include "../decay.h"
-#include "../globals/settings.h"
-#include "../globals/uvars.h"
-#include "../item/item.h"
-#include "../item/itemdesc.h"
-#include "../listenpt.h"
-#include "../packetscrobj.h"
-#include "../polcfg.h"
-#include "../proplist.h"
-#include "../realms.h"
-#include "../realms/realm.h"
-#include "../tooltips.h"
-#include "../uobject.h"
-#include "../uoexec.h"
+#include "bscript/berror.h"
+#include "bscript/bobject.h"
+#include "bscript/dict.h"
+#include "bscript/impstr.h"
+#include "clib/clib.h"
+#include "clib/clib_MD5.h"
+#include "clib/fileutil.h"
+#include "clib/rawtypes.h"
+#include "clib/strutil.h"
+#include "clib/threadhelp.h"
+#include "plib/pkg.h"
+#include "plib/systemstate.h"
+
+
+#include "realms/realms.h"
+#include "realms/realm.h"
+
+#include "globals/settings.h"
+#include "globals/uvars.h"
+#include "item/item.h"
+#include "item/itemdesc.h"
+
+#include "cmdlevel.h"
+#include "core.h"
+#include "decay.h"
+#include "listenpt.h"
+#include "packetscrobj.h"
+#include "polcfg.h"
+#include "proplist.h"
+#include "tooltips.h"
+#include "uobject.h"
+#include "uoexec.h"
 
 #include <module_defs/polsys.h>
