+++ pol-core/pol/module/storagemod.cpp
@@ -13,7 +13,7 @@
 #include "../../clib/clib.h"
 #include "../globals/uvars.h"
 #include "../item/item.h"
-#include "../realms.h"
+#include "realms/realms.h"
 #include "../storage.h"
 #include "../uoscrobj.h"
