+++ pol-core/pol/module/uomod.cpp
@@ -135,8 +135,8 @@
 #include "../polobject.h"
 #include "../polsig.h"
 #include "../profile.h"
-#include "../realms.h"
-#include "../realms/realm.h"
+#include "realms/realms.h"
+#include "realms/realm.h"
 #include "../resource.h"
 #include "../savedata.h"
 #include "../scrdef.h"
