+++ pol-core/pol/module/uomod4.cpp
@@ -24,8 +24,8 @@
 #include "../mobile/charactr.h"
 #include "../multi/boat.h"
 #include "../polclass.h"
-#include "../realms.h"
-#include "../realms/realm.h"
+#include "realms/realms.h"
+#include "realms/realm.h"
 #include "../ufunc.h"
 #include "../uobject.h"
 #include "../uworld.h"
