+++ pol-core/pol/uobject.cpp
@@ -31,7 +31,7 @@
 #include "objtype.h"
 #include "polcfg.h"
 #include "proplist.h"
-#include "realms.h"
+#include "realms/realms.h"
 #include "realms/realm.h"
 #include "syshookscript.h"
 #include "tooltips.h"
